+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## 0.20.1 - 2024-08-26
+
+### Enhancements
+- Added `DynAsyncBufWriter` for buffering compressed or uncompressed async output
+- Added new publisher values for `XCIS.BBOTRADES` and `XNYS.BBOTRADES`
+
+### Bug fixes
+- Added missing Python type stub for `pretty_ts_ref` in `StatMsg`
+
 ## 0.20.0 - 2024-07-30
 
 ### Enhancements